Update lamp it

+++ src/main/java/com/banhodepote/api/repository/WaiterRepository.java
@@ -6,8 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.banhodepote.api.model.Waiter;
 
-public interface WaiterRepository extends JpaRepository<Waiter, Long> {
+public interface WaiterRepository extends JpaRepository<Waiter, Integer> {
     
     List<Waiter> findAllByName(String name);
+
 
 }
